PROYECTO TERMINADO

+++ database/migrations/2019_08_01_033051_create_tipo_creditos_table.php
@@ -17,7 +17,7 @@ class CreateTipoCreditosTable extends Migration
             $table->bigIncrements('id');
             $table->string('nombre');
             $table->string('descripcion');
-            $table->double('interes_fijo');
+            $table->float('interes_fijo');
             $table->timestamps();
         });
     }
